Moved some commands around to clean them in different folders
