+++ src/main/java/me/glaremasters/guilds/commands/member/CommandDecline.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package me.glaremasters.guilds.commands;
+package me.glaremasters.guilds.commands.member;
 
 import co.aikar.commands.ACFUtil;
 import co.aikar.commands.BaseCommand;
@@ -36,63 +36,47 @@ import co.aikar.commands.annotation.Syntax;
 import co.aikar.commands.annotation.Values;
 import lombok.AllArgsConstructor;
 import me.glaremasters.guilds.messages.Messages;
-import me.glaremasters.guilds.api.events.GuildJoinEvent;
 import me.glaremasters.guilds.exceptions.ExpectationNotMet;
 import me.glaremasters.guilds.guild.Guild;
 import me.glaremasters.guilds.guild.GuildHandler;
-import me.glaremasters.guilds.guild.GuildMember;
 import me.glaremasters.guilds.utils.Constants;
-import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
 /**
  * Created by Glare
  * Date: 4/5/2019
- * Time: 10:29 PM
+ * Time: 10:25 PM
  */
 @AllArgsConstructor @CommandAlias(Constants.ROOT_ALIAS)
-public class CommandAccept extends BaseCommand {
+public class CommandDecline extends BaseCommand {
 
     private GuildHandler guildHandler;
 
     /**
-     * Accept a guild invite
-     * @param player the player accepting the invite
-     * @param name the name of the guild being accepted
+     * Decline a guild invite
+     * @param player the player declining the invite
+     * @param name the name of the guild
      */
-    @Subcommand("accept|join")
-    @Description("{@@descriptions.accept}")
-    @CommandPermission(Constants.BASE_PERM + "accept")
+    @Subcommand("decline")
+    @Description("{@@descriptions.decline}")
+    @CommandPermission(Constants.BASE_PERM + "decline")
     @CommandCompletion("@invitedTo")
     @Syntax("<guild name>")
     public void execute(Player player, @Values("@invitedTo") @Single String name) {
-        if (guildHandler.getGuild(name) != null)
-            ACFUtil.sneaky(new ExpectationNotMet(Messages.ERROR__ALREADY_IN_GUILD));
-
         Guild guild = guildHandler.getGuild(name);
 
         if (guild == null)
             ACFUtil.sneaky(new ExpectationNotMet(Messages.ERROR__GUILD_NO_EXIST));
 
-        if (!guild.checkIfInvited(player) && guild.isPrivate())
+        if (guildHandler.getGuild(player) != null)
+            ACFUtil.sneaky(new ExpectationNotMet(Messages.ERROR__ALREADY_IN_GUILD));
+
+        if (!guild.checkIfInvited(player))
             ACFUtil.sneaky(new ExpectationNotMet(Messages.ACCEPT__NOT_INVITED));
 
-        if (guild.checkIfFull())
-            ACFUtil.sneaky(new ExpectationNotMet(Messages.ACCEPT__GUILD_FULL));
+        guild.removeInvitedMember(player.getUniqueId());
 
-        GuildJoinEvent event = new GuildJoinEvent(player, guild);
-        Bukkit.getPluginManager().callEvent(event);
-
-        if (event.isCancelled())
-            return;
-
-        guild.addMember(new GuildMember(player.getUniqueId(), guildHandler.getLowestGuildRole()));
-
-        guild.sendMessage(getCurrentCommandManager(), Messages.ACCEPT__PLAYER_JOINED,
-                "{player}", player.getName());
-
-        getCurrentCommandIssuer().sendInfo(Messages.ACCEPT__SUCCESSFUL,
-                "{guild}", guild.getName());
+        getCurrentCommandIssuer().sendInfo(Messages.DECLINE__SUCCESS);
     }
 
 }
